fix tests

+++ controllers/cards.js
@@ -1,3 +1,4 @@
+const mongoose = require('mongoose');
 const Card = require('../models/card');
 const { NOT_FOUND, BAD_REQUEST, INTERNAL_ERROR } = require('./consts');
 
@@ -26,58 +27,65 @@ module.exports.createCard = (req, res) => {
 
 // удаляет карточку по идентификатору
 module.exports.removeCard = (req, res) => {
+  if (!mongoose.isValidObjectId(req.params.cardsId)) {
+    res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные id' });
+    return;
+  }
   Card.findByIdAndRemove(req.params.cardsId)
-    .then((card) => res.send({ data: card }))
-    .catch((err) => {
-      console.log(err.name);
-
-      if (err.name === 'CastError') {
+    .then((card) => {
+      if (card === null) {
         res.status(NOT_FOUND).send({ message: 'Передан несуществующий _id карточки' });
-      } else if (err.name === 'ValidationError') {
-        res.status(BAD_REQUEST).send({ message: 'Передан некорректный _id карточки' });
-      } else {
-        res.status(INTERNAL_ERROR).send({ message: 'Произошла ошибка' });
+        return;
       }
+      res.send({ data: card });
+    })
+    .catch((err) => {
+      res.status(INTERNAL_ERROR).send({ message: 'Произошла ошибка' });
     });
 };
 
 // поставить лайк карточке
 module.exports.putLikeCard = (req, res) => {
+  if (!mongoose.isValidObjectId(req.params.cardsId)) {
+    res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные id для постановки лайка' });
+    return;
+  }
   Card.findByIdAndUpdate(
     req.params.cardsId,
     { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
     { new: true },
   )
     .then((card) => {
-      console.log(card);
+      if (card === null) {
+        res.status(NOT_FOUND).send({ message: 'Передан несуществующий _id карточки' });
+        return;
+      }
       res.send({ data: card });
     })
     .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные для постановки лайка' });
-      } else if (err.name === 'ValidationError') {
-        res.status(NOT_FOUND).send({ message: 'Передан несуществующий _id карточки' });
-      } else {
-        res.status(INTERNAL_ERROR).send({ message: 'Произошла ошибка' });
-      }
+      res.status(INTERNAL_ERROR).send({ message: 'Произошла ошибка' });
     });
 };
 
 // убрать лайк с карточки
 module.exports.deleteLikeCard = (req, res) => {
+  if (!mongoose.isValidObjectId(req.params.cardsId)) {
+    res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные id для постановки лайка' });
+    return;
+  }
   Card.findByIdAndUpdate(
     req.params.cardsId,
     { $pull: { likes: req.user._id } }, // убрать _id из массива
     { new: true },
   )
-    .then((card) => res.send({ data: card }))
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные для снятия лайка' });
-      } else if (err.name === 'ValidationError') {
+    .then((card) => {
+      if (card === null) {
         res.status(NOT_FOUND).send({ message: 'Передан несуществующий _id карточки' });
-      } else {
-        res.status(INTERNAL_ERROR).send({ message: 'Произошла ошибка' });
+        return;
       }
+      res.send({ data: card });
+    })
+    .catch((err) => {
+      res.status(INTERNAL_ERROR).send({ message: 'Произошла ошибка' });
     });
 };
